Refactor component and test folders for easy navigation
