+++ src/components/App/App.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import renderer from 'react-test-renderer';
-import App from '../components/App';
+import App from '.';
 
 describe('App', () => {
   it('renders without crashing', () => {
