+++ src/components/Search/Search.test.js
@@ -1,18 +1,18 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import renderer from 'react-test-renderer';
-import Button from '../components/Button';
+import Search from '.';
 
-describe('Button', () => {
+describe('Search', () => {
   it('renders without crashing', () => {
     const div = document.createElement('div');
-    ReactDOM.render(<Button>Give Me More</Button>, div);
+    ReactDOM.render(<Search>Search</Search>, div);
     ReactDOM.unmountComponentAtNode(div);
   });
 
   test('has a valid snapshot', () => {
     const component = renderer.create(
-      <Button>Give Me More</Button>
+      <Search>Search</Search>
     );
     let tree = component.toJSON();
     expect(tree).toMatchSnapshot();
