+++ src/components/Table/Table.test.js
@@ -3,7 +3,8 @@ import ReactDOM from 'react-dom';
 import renderer from 'react-test-renderer';
 import Enzyme, { shallow } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
-import Table from '../components/Table/index';
+
+import Table from '.';
 
 Enzyme.configure({ adapter: new Adapter() });
 
@@ -18,6 +19,7 @@ describe('Table', () => {
   it('renders without crashing', () => {
     const div = document.createElement('div');
     ReactDOM.render(<Table {...props} />, div);
+    ReactDOM.unmountComponentAtNode(div);
   });
 
   test('has a valid snapshot', () => {
