+++ src/utils/constants.js
@@ -1,0 +1,11 @@
+const api_url = {
+  DEFAULT_QUERY : 'redux',
+  DEFAULT_HPP : '100',
+  PATH_BASE : 'https://hn.algolia.com/api/v1',
+  PATH_SEARCH : '/search',
+  PARAM_SEARCH : 'query=',
+  PARAM_PAGE : 'page=',
+  PARAM_HPP: 'hitsPerPage='
+}
+
+export default api_url;
