+++ src/utils/index.js
@@ -1,0 +1,5 @@
+//Entry point into utils
+
+import api_url from './constants';
+
+export default api_url;
